Check profile path before the removal

+++ framework/python/src/core/session.py
@@ -715,7 +715,12 @@ class TestrunSession():
         old_name = profile_json.get('name')
 
         # Delete the original file
-        os.remove(os.path.join(PROFILES_DIR, old_name + '.json'))
+        old_file_path = os.path.join(PROFILES_DIR, old_name + '.json')
+        if os.path.exists(old_file_path):
+          try:
+            os.remove(old_file_path)
+          except OSError as e:
+            LOGGER.error(f'An error occurred whilst deleting old profile file {old_file_path}: {e}')
 
     # Write file to disk
     with open(os.path.join(PROFILES_DIR, risk_profile.name + '.json'),
@@ -865,7 +870,8 @@ question {question.get('question')}''')
 
       profile_path = os.path.join(PROFILES_DIR, file_name)
 
-      os.remove(profile_path)
+      if os.path.exists(profile_path):
+        os.remove(profile_path)
       self._profiles.remove(profile)
 
       return True
